refactor: added __init__, imported Netlist in it
