+++ openfasoc/generators/gdsfactory-gen/glayout/spice/__init__.py
@@ -1,0 +1,1 @@
+from .netlist import Netlist
